atomicディレクトリ構成にする
