+++ package.json
@@ -19,6 +19,7 @@
     "react-paginate": "^6.5.0",
     "react-slick": "^0.27.12",
     "react-transition-group": "^4.4.1",
+    "sass": "^1.32.8",
     "slick-carousel": "^1.8.1"
   },
   "devDependencies": {
