--- src/components/Button.tsx
@@ -1,36 +1,0 @@
-import React from 'react'
-import Link from 'next/link';
-import styles from '../components-style/Button.module.css';
-
-interface Props {
-  text: string;
-  path?: string;
-}
-
-const Button = (props: Props) => {
-  return (
-    <Link href={props.path}>
-      <a href="/" className={styles.btn}>
-      <svg width="277" height="62">
-        <defs>
-            <linearGradient id="grad1">
-                {/* <stop offset="0%" stop-color="#FF8282"/>
-                <stop offset="100%" stop-color="#E178ED" /> */}
-                <stop offset="0%" stopColor="#0398c3"/>
-                <stop offset="25%" stopColor="#0398c3"/>
-                <stop offset="25%" stopColor="#ff0202"/>
-                <stop offset="50%" stopColor="#ff0202"/>
-                <stop offset="75%" stopColor="#ff0202"/>
-                <stop offset="75%" stopColor="#03ae65" />
-                <stop offset="100%" stopColor="#03ae65" />
-            </linearGradient>
-        </defs>
-        <rect x="5" y="5" rx="25" fill="none" stroke="url(#grad1)" width="266" height="50"></rect>
-      </svg>
-      <span>{props.text}</span>
-      </a>
-    </Link>
-  )
-}
-
-export default Button
