+++ src/components/index.ts
@@ -6,7 +6,7 @@ export {default as Slick} from './Slick';
 export {default as PostThumbnail} from './PostThumbnail';
 export {default as ContentIndex} from './ContentIndex';
 export {default as PostFlex} from './PostFlex';
-export {default as Button} from './Button';
+export {default as Button} from './atom/Button/Button';
 export {default as CategoryArea} from './CategoryArea';
 export {default as LangToggler} from './LangToggler';
 export {default as LangToggler2} from './LangToggler2';
